Updating

+++ src/test/java/BaseClass/BaseClass.java
@@ -1,30 +1,42 @@
 package BaseClass;
 
 
+import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
+import java.text.SimpleDateFormat;
 import java.time.Duration;
+import java.util.Date;
 import java.util.Properties;
 
 import org.apache.commons.lang3.RandomStringUtils;
-
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+
+
+import org.apache.logging.log4j.LogManager;  //Log4j
+import org.apache.logging.log4j.Logger;  //Log4j
+
 
 public class BaseClass {
 	public WebDriver driver;
 	public Logger logger;
 	public Properties p;
 	
-	@BeforeClass
+	@BeforeClass(groups= {"Sanity","Regression", "Master", "Datadriven"})
 	@Parameters({"OS","browser"})
-	public void setup(String OS, String br) throws IOException {
+	public void setup(String os, String br) throws IOException {
 		logger=LogManager.getLogger(this.getClass());
 		
 		//Load property file
@@ -32,11 +44,59 @@ public class BaseClass {
 		p=new Properties();
 		p.load(fd);
 		
-		switch (br.toLowerCase()) {
-		case "chrome" : driver=new ChromeDriver(); break;
-		case "edge" : driver=new EdgeDriver(); break;
-		default : System.out.println("Invald Browser"); return;
+		
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote"))
+		{
+			DesiredCapabilities capabilities=new DesiredCapabilities();
+			
+			//os
+			if(os.equalsIgnoreCase("windows"))
+			{
+				capabilities.setPlatform(Platform.WIN11);
+			}
+			else if(os.equalsIgnoreCase("linux"))
+			{
+				capabilities.setPlatform(Platform.LINUX);
+				
+			}
+			else if (os.equalsIgnoreCase("mac"))
+			{
+				capabilities.setPlatform(Platform.MAC);
+			}
+			else
+			{
+				System.out.println("No matching os");
+				return;
+			}
+			
+			//browser
+			switch(br.toLowerCase())
+			{
+			case "chrome": capabilities.setBrowserName("chrome"); break;
+			case "edge": capabilities.setBrowserName("MicrosoftEdge"); break;
+			case "firefox": capabilities.setBrowserName("firefox"); break;
+			default: System.out.println("No matching browser"); return;
+			}
+			
+			driver=new RemoteWebDriver(new URL("http://localhost:4444/wd/hub"),capabilities);
 		}
+		
+				
+		if(p.getProperty("execution_env").equalsIgnoreCase("local"))
+		{
+
+			switch(br.toLowerCase())
+			{
+			case "chrome" : driver=new ChromeDriver(); break;
+			case "edge" : driver=new EdgeDriver(); break;
+			case "firefox": driver=new FirefoxDriver(); break;
+			default : System.out.println("Invalid browser name.."); return;
+			}
+		}
+
+
+		
+		
 		
 		driver.manage().deleteAllCookies();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(3));
@@ -45,7 +105,7 @@ public class BaseClass {
 		driver.manage().window().maximize();
 	}
 	
-	@AfterClass
+	@AfterClass(groups= {"Sanity","Regression", "Master", "Datadriven"})
 	public void tearDown() {
 		driver.quit();
 	}
@@ -62,5 +122,21 @@ public class BaseClass {
 		String genRanChar = RandomStringUtils.randomAlphabetic(3);
 		String genRanNo = RandomStringUtils.randomNumeric(7);
 		return (genRanChar+"@"+genRanNo);	
+	}
+	
+	public String captureScreen(String tname) throws IOException {
+
+		String timeStamp = new SimpleDateFormat("yyyyMMddhhmmss").format(new Date());
+				
+		TakesScreenshot takesScreenshot = (TakesScreenshot) driver;
+		File sourceFile = takesScreenshot.getScreenshotAs(OutputType.FILE);
+		
+		String targetFilePath=System.getProperty("user.dir")+"\\screenshots\\" + tname + "_" + timeStamp + ".png";
+		File targetFile=new File(targetFilePath);
+		
+		sourceFile.renameTo(targetFile);
+			
+		return targetFilePath;
+
 	}
 }
